requirements.txt
requests
python-dotenv
beautifulsoup4
openai
flask
flask-cors
